+++ console/folder/migrate/migration/1474812248_city.php
@@ -5,6 +5,6 @@ $q = array();
 $q[] = 'CREATE TABLE `city`
         (
             `city_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `city_country_id` SMALLINT(3) DEFAULT 0,
+            `city_country_id` INT(3) DEFAULT 0,
             `city_name` VARCHAR(255) NOT NULL
         );';
