+++ console/folder/migrate/migration/1474997831_surnamecountry.php
@@ -4,6 +4,6 @@ $q = array();
 
 $q[] = 'CREATE TABLE `surnamecountry`
         (
-            `surnamecountry_country_id` SMALLINT(3) DEFAULT 0,
+            `surnamecountry_country_id` INT(3) DEFAULT 0,
             `surnamecountry_surname_id` INT(11) DEFAULT 0
         );';
