+++ console/folder/migrate/migration/1474997837_namecountry.php
@@ -4,6 +4,6 @@ $q = array();
 
 $q[] = 'CREATE TABLE `namecountry`
         (
-            `namecountry_country_id` SMALLINT(3) DEFAULT 0,
+            `namecountry_country_id` INT(3) DEFAULT 0,
             `namecountry_name_id` INT(11) DEFAULT 0
         );';
