+++ console/folder/migrate/migration/1475587849_playerposition.php
@@ -5,5 +5,5 @@ $q = array();
 $q[] = 'CREATE TABLE `playerposition`
         (
             `playerposition_player_id` INT(11) DEFAULT 0,
-            `playerposition_position_id` TINYINT(1) DEFAULT 0
+            `playerposition_position_id` INT(1) DEFAULT 0
         );';
