+++ console/folder/migrate/migration/1477633883_division.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `division`
         (
-            `division_id` TINYINT(2) PRIMARY KEY AUTO_INCREMENT,
+            `division_id` INT(2) PRIMARY KEY AUTO_INCREMENT,
             `division_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `division` (`division_name`)
