+++ console/folder/migrate/migration/1479108376_lineup.php
@@ -5,19 +5,19 @@ $q = array();
 $q[] = 'CREATE TABLE `lineup`
         (
             `lineup_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `lineup_age` TINYINT(2) DEFAULT 0,
-            `lineup_assist` TINYINT(3) DEFAULT 0,
+            `lineup_age` INT(2) DEFAULT 0,
+            `lineup_assist` INT(3) DEFAULT 0,
             `lineup_game_id` INT(1) DEFAULT 0,
-            `lineup_line_id` TINYINT(1) DEFAULT 0,
-            `lineup_national_id` SMALLINT(5) DEFAULT 0,
-            `lineup_pass` TINYINT(3) DEFAULT 0,
-            `lineup_penalty` TINYINT(3) DEFAULT 0,
+            `lineup_line_id` INT(1) DEFAULT 0,
+            `lineup_national_id` INT(5) DEFAULT 0,
+            `lineup_pass` INT(3) DEFAULT 0,
+            `lineup_penalty` INT(3) DEFAULT 0,
             `lineup_player_id` INT(1) DEFAULT 0,
-            `lineup_plus_minus` TINYINT(3) DEFAULT 0,
-            `lineup_position_id` TINYINT(1) DEFAULT 0,
-            `lineup_power_nominal` SMALLINT(3) DEFAULT 0,
-            `lineup_power_real` SMALLINT(3) DEFAULT 0,
-            `lineup_score` TINYINT(3) DEFAULT 0,
-            `lineup_shot` TINYINT(3) DEFAULT 0,
-            `lineup_team_id` SMALLINT(5) DEFAULT 0
+            `lineup_plus_minus` INT(3) DEFAULT 0,
+            `lineup_position_id` INT(1) DEFAULT 0,
+            `lineup_power_nominal` INT(3) DEFAULT 0,
+            `lineup_power_real` INT(3) DEFAULT 0,
+            `lineup_score` INT(3) DEFAULT 0,
+            `lineup_shot` INT(3) DEFAULT 0,
+            `lineup_team_id` INT(5) DEFAULT 0
         );';
