+++ console/folder/migrate/migration/1479193028_mood.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `mood`
         (
-            `mood_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
+            `mood_id` INT(1) PRIMARY KEY AUTO_INCREMENT,
             `mood_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `mood` (`mood_name`)
