+++ console/folder/migrate/migration/1480011637_constructiontype.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `construnctiontype`
         (
-            `construnctiontype_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
+            `construnctiontype_id` INT(1) PRIMARY KEY AUTO_INCREMENT,
             `construnctiontype_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `construnctiontype` (`construnctiontype_name`)
