+++ console/folder/migrate/migration/1480243504_buildingstadium.php
@@ -5,9 +5,9 @@ $q = array();
 $q[] = 'CREATE TABLE `buildingstadium`
         (
             `buildingstadium_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `buildingstadium_capacity` SMALLINT(5) DEFAULT 0,
-            `buildingstadium_constructiontype_id` TINYINT(1) DEFAULT 0,
-            `buildingstadium_day` TINYINT(2) DEFAULT 0,
-            `buildingstadium_ready` TINYINT(1) DEFAULT 0,
-            `buildingstadium_team_id` SMALLINT(5) DEFAULT 0
+            `buildingstadium_capacity` INT(5) DEFAULT 0,
+            `buildingstadium_constructiontype_id` INT(1) DEFAULT 0,
+            `buildingstadium_day` INT(2) DEFAULT 0,
+            `buildingstadium_ready` INT(1) DEFAULT 0,
+            `buildingstadium_team_id` INT(5) DEFAULT 0
         );';
